+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -148,6 +148,7 @@ public let masterRuleList = RuleList(rules: [
     ValidIBInspectableRule.self,
     VerticalParameterAlignmentOnCallRule.self,
     VerticalParameterAlignmentRule.self,
+    VerticalWhitespaceBetweenCasesRule.self,
     VerticalWhitespaceRule.self,
     VoidReturnRule.self,
     WeakDelegateRule.self,